Refactored (incompletely) BoardWidget by extracting non-GUI functionality to the new Game class. Switched UCIEngine and co to use Qt's signal/slot mechanism instead of manual callbacks. Added engine id (name), which is distinct from the UCI name. Introduced concept of engine presets.

+++ QtChessGUI/Core/misc.cpp
@@ -2,14 +2,14 @@
 
 std::string misc::trim_left(std::string str, const std::string& chars)
 {
-	if (const int beg = str.find_first_not_of(chars); beg != std::string::npos)
+	if (const size_t beg = str.find_first_not_of(chars); beg != std::string::npos)
 		str.erase(0, beg);
 	return str;
 }
 
 std::string misc::trim_right(std::string str, const std::string& chars)
 {
-	if (const int last = str.find_last_not_of(chars); last != std::string::npos)
+	if (const size_t last = str.find_last_not_of(chars); last != std::string::npos)
 		str.erase(last + 1);
 	return str;
 }
